requirements.txt
asgiref==3.4.1
django==3.2.5
django-haystack==3.1.1
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-select2==7.7.2
django-taggit==1.5.1
django-redis==5.0.0
django-heroku==0.3.1
dj_database_url==0.5.0
gunicorn==20.1.0
markdown==3.3.4
numpy==1.21.2
Pillow==8.3.2
pysolr==3.9.0
redis==3.5.3
social-auth-app-django==5.0.0
whitenoise==5.3.0
psycopg2==2.9.1
python-decouple==3.4